+++ Graph store/Hash/HashFunction.h
@@ -7,13 +7,13 @@ template <typename Key>
 class HashFunction
 {
 public:
-	size_t operator()(const Key&)const;
+	unsigned operator()(const Key&)const;
 };
 
-
 template <typename Key>
-inline size_t HashFunction<Key>::operator()(const Key& key)const
+inline unsigned HashFunction<Key>::operator()(const Key& key)const
 {
-	return static_cast<size_t>(key);
+	return static_cast<unsigned>(key);
 }
+
 #endif //__HASH_FUNC_H_INCLUDED__
